requirements.txt
ultralytics==8.3.88
roboflow==1.1.56
matplotlib==3.10.1
opencv-python==4.11.0.86
tqdm==4.67.1
pandas==2.2.3
face-recognition==1.3.0
numpy==2.1.1
dlib==19.24.6
pillow==11.0.0
requests==2.32.3
pyarrow==19.0.1
fastparquet==2024.11.0
wandb==0.19.8
PyYAML==6.0.2